Implement registration page with admin verification and form factory for consistent styling
