+++ app/ui/pages/__init__.py
@@ -9,3 +9,4 @@ from .maintenance_page import MaintenancePage
 from .invoice.invoice_page import InvoicePage
 from .help_page import HelpPage
 from .about_page import AboutPage
+from .register.register_page import RegisterPage
