+++ app/ui/pages/register/__init__.py
@@ -1,0 +1,1 @@
+from .register_page import RegisterPage
