+++ app/ui/pages/register/dialogs/__init__.py
@@ -1,0 +1,1 @@
+from .admin_verification_dialog import AdminVerificationDialog
